requirements.txt
## Core FastAPI microservices stack
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
starlette>=0.27.0
redis>=5.0.0
aiohttp>=3.9.0
python-multipart>=0.0.6
pyjwt

# Azure services - Linux App Service compatible versions
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-cognitiveservices-speech>=1.45.0
azure-communication-callautomation>=1.4.0
azure-communication-identity>=1.5.0
azure-keyvault-secrets>=4.7.0
azure-eventgrid>=4.10.0

# Azure monitoring and telemetry
azure-monitor-opentelemetry>=1.6.11
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-openai-v2

# Azure + misc
azure-core>=1.29.0
colorama>=0.4.6

# AI and OpenAI integration
openai>=1.50.0

# HTTP clients - stable versions
httpx>=0.27.0
aiofiles>=23.0.0

# Async and networking tools
tenacity>=8.5.0
# Load testing (moved from end)
locust>=2.20.0
# WebSocket and communication libraries  
websockets>=12.0
websocket-client>=1.6.0
pytest>=7.4.0
backoff>=2.0.0

# Data processing and YAML configuration
numpy>=1.24.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0
jinja2>=3.1.0
typing-extensions>=4.8.0
langdetect>=1.0.9
PyYAML>=6.0.0
pyyaml-include>=1.3.0

# Audio processing for Live Voice API and samples
sounddevice>=0.4.6
pyaudio>=0.2.11
wave  # Part of standard library, but may need explicit declaration

# Database drivers
pymongo>=4.6.0
rapidfuzz>=3.13.0